Cache context is declared via ::getCacheContexts()
We had to go back to implementing block::getCacheContexts().
We need to know the cache context when loading the block;
not waiting until having rendered it.

Also, do not confuse the "Cache Context" with the "Context".
The block's context refers to what data is available, and is
declared via the annotation.

Updates #8